Module Autowp.Logic

type term =

| Int of int

| Var of string

| Fun of string * term list

Terms of the logic

type formula =

| Forall of string list * formula

| Exitsts of string list * formula

| Pred of string * term list

| And of formula * formula

| Or of formula * formula

| Impl of formula * formula

| Not of formula

Formulae of the logic

val subst_term : string -> term -> term -> term

Substitute an id with a term in a term

val subst : string -> term -> formula -> formula

Substitute an id with a term in a formula

val commas : string list -> string

val str_of_term : term -> string

Term to string conversion

val priority : formula -> int

val str_of_form : formula -> string

Formula to string conversion