Module Autowp.Parser

parsing helpers

val some : ('a -> ('b * 'a) option) -> 'a -> ('b list * 'a) option

Parse one or more ps

val let* : ('a -> ('b * 'c) option) -> ('b -> 'c -> 'd option) -> 'a -> 'd option

Let notation for bindings

val parens : (char Opal.LazyStream.t -> ('a * char Opal.LazyStream.t) option) -> char Opal.LazyStream.t -> ('a * char Opal.LazyStream.t) option

Usefull parsers

val parse_id : char Opal.LazyStream.t -> (string * char Opal.LazyStream.t) option

Parse an identifier

val parse_nat : char Opal.LazyStream.t -> (int * char Opal.LazyStream.t) option

Parse a (non negative) natural number

val parse_rel : char Opal.LazyStream.t -> (int * char Opal.LazyStream.t) option

Parse a relative number

val parse_int : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse an integer number

val parse_var : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse variable

val parse_commasep : (char Opal.LazyStream.t -> ('a * char Opal.LazyStream.t) option) -> char Opal.LazyStream.t -> ('a list * char Opal.LazyStream.t) option

Parse items separated by a comma

val parse_args : (char Opal.LazyStream.t -> ('a * char Opal.LazyStream.t) option) -> char Opal.LazyStream.t -> ('a list * char Opal.LazyStream.t) option

Parse items separated by a comma and parenthesized

Parse arithmetic operators

val op_add : char Opal.LazyStream.t -> ((Logic.term -> Logic.term -> Logic.term) * char Opal.LazyStream.t) option

val op_sub : char Opal.LazyStream.t -> ((Logic.term -> Logic.term -> Logic.term) * char Opal.LazyStream.t) option

val op_mul : char Opal.LazyStream.t -> ((Logic.term -> Logic.term -> Logic.term) * char Opal.LazyStream.t) option

val op_div : char Opal.LazyStream.t -> ((Logic.term -> Logic.term -> Logic.term) * char Opal.LazyStream.t) option

Parsers for logic connectors

val op_and : char Opal.LazyStream.t -> ((Logic.formula -> Logic.formula -> Logic.formula) * char Opal.LazyStream.t) option

val op_or : char Opal.LazyStream.t -> ((Logic.formula -> Logic.formula -> Logic.formula) * char Opal.LazyStream.t) option

val op_impl : char Opal.LazyStream.t -> ((Logic.formula -> Logic.formula -> Logic.formula) * char Opal.LazyStream.t) option

Parsers for arithmetic expressions

val parse_expr : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse an expression

val parse_product : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse a product

val parse_negation : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse the negation of an expression

val parse_factor : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse a factor. A factor is either a negation, a parenthesised expression, a variable or an integer

Parsers for arithmetic comparisons

val comparators : (char Opal.LazyStream.t -> (string * char Opal.LazyStream.t) option) list

A list of parsers for comparison symbols (<=, >=, <, >, =)

val parse_comp : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a comparison expression

Parsers for logic expressions

val parse_pred : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a predicate

val parse_fun : char Opal.LazyStream.t -> (Logic.term * char Opal.LazyStream.t) option

Parse a function application

val parse_form : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a logic formula

val parse_forall : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a formula starting with a universal quantificator

val parse_exists : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a formula starting with an existential quantificator

val parse_nq_form : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a formula which does'nt start with a qantificator

val parse_disjunction : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a disjunction

val parse_conjunction : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a conjunction

val parse_not : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse a negation

val parse_atom : char Opal.LazyStream.t -> (Logic.formula * char Opal.LazyStream.t) option

Parse an atomic formula. An atomic formula is either a quantified formula, a comparison, a negation, a single predicate applied to arguments or a parenthesized formula

Parsers for IMP programs

val parse_seq : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse a sequence of statements

val parse_stmt : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse a single statement

val parse_aff : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse an affectation

val parse_if : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse a conditional (withou Else branch)

val parse_ifElse : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse a conditional

val parse_while : char Opal.LazyStream.t -> (Imp.stmt * char Opal.LazyStream.t) option

Parse while loop

String parsers

val parse_spec : string -> Logic.formula

val parse_prog : string -> Imp.stmt