Module Autowp__Imp

A minimalist imperative programming language to write small arithmetic algorithms.

To simplify the computation of the Weakest Preconditions, we assume that expressions (resp. conditions) in IMP are equivalent to terms (resp. formulae) in Logic.

type expr = Autowp.Logic.term

Expressions of IMP are terms in Logic

type cond = Autowp.Logic.formula

Conditions of IMP are formulas in Logic

type invariant = Autowp.Logic.formula

Invariants are formulas in Logic

type variant = expr

Variants are formulas in Logic

type stmt =

| Aff of string * expr

| If of cond * stmt

| IfElse of cond * stmt * stmt

| While of invariant * variant * cond * stmt

| Seqc of stmt * stmt

IMP statements TODO : while loop

module VarSet : sig ... end

val vars : stmt -> VarSet.elt list

val wp : stmt -> Autowp.Logic.formula -> Autowp.Logic.formula

wp prog post computes the weakest precondition of program prog with respect to a postcondition post. wp computes preconditions for partial correction only.

TODO : Add variants and terminations

val make_decr : Autowp.Logic.formula -> Autowp.Logic.term -> Autowp.Logic.formula -> stmt -> Autowp.Logic.formula

val seqc_of_list : stmt list -> stmt

Convert a list of statements into a sequence